Split out Phrase context
