+++ fonduer/parser/models/__init__.py
@@ -1,12 +1,8 @@
-from fonduer.parser.models.context import (
-    Context,
-    Phrase,
-    construct_stable_id,
-    split_stable_id,
-)
+from fonduer.parser.models.context import Context, construct_stable_id, split_stable_id
 from fonduer.parser.models.document import Document
 from fonduer.parser.models.figure import Figure
 from fonduer.parser.models.image import Image, TemporaryImage
+from fonduer.parser.models.phrase import Phrase
 from fonduer.parser.models.span import (
     ImplicitSpan,
     Span,
